updates to brave helpers

+++ brave/brave.zsh
@@ -9,22 +9,71 @@ export SCCACHE_DIR=$HOME/sccache
 if [[ "$OS_NAME" = "Darwin" ]]; then
     BRAVE_RAMDISK_ENABLED=$((`hostinfo | grep 'Primary memory available' | sed 's/[a-zA-Z ]*: //' | sed 's/ gigabytes//'` > 128))
     export BRAVE_NIGHTLY="/Applications/Brave Browser Nightly.app/Contents/MacOS/Brave Browser Nightly"
-    if [[ $BRAVE_RAMDISK_ENABLED ]]; then
+    if [[ "$BRAVE_RAMDISK_ENABLED" = "1" ]]; then
         export SCCACHE_DIR=$BRAVE_RAMDISK/sccache
     fi
 fi
+
+brave_profile() {
+    brave_bin=$1
+    shift
+
+    if [ ! -x "$brave_bin" ]; then
+        echo "Invalid Brave binary: $brave_bin"
+        return 1
+    fi
+
+    profile=$1
+    shift
+
+    if [ -z "$profile" ]; then
+        echo "Missing profile dir"
+        return 1
+    fi
+
+    "$brave_bin" --user-data-dir="$profile" "$@"
+}
+
+brave_tmp_profile() {
+    brave_bin=$1
+    shift
+
+    if [ ! -x "$brave_bin" ]; then
+        echo "Invalid Brave binary: $brave_bin"
+        return 1
+    fi
+
+    tmpdir=$(mktemp -d "/tmp/brave.XXXXXXX")
+    brave_profile "$brave_bin" "$tmpdir" "$@"
+
+    if [[ -d "$tmpdir" ]]; then
+        rm -rf "$tmpdir"
+    fi
+}
 
 brave_dev() {
     npm start "$@"
 }
 
 brave_dev_clean() {
-    tmpdir=$(mktemp -d "/tmp/brave.XXXXXXX")
+    brave_tmp_profile brave_dev
+}
 
-    brave_dev -- --user-data-dir="$tmpdir" "$@"
-    if [[ -d "$tmpdir" ]]; then
-        rm -rf "$tmpdir"
+brave_dev_dev_wallet() {
+    brave_bin="$PWD/src/out/Component/Brave Browser Development.app/Contents/MacOS/Brave Browser Development"
+    if [[ ! -x "$brave_bin" ]]; then
+        echo "No build of brave found at $brave_bin"
+        return 1
     fi
+
+    wallet=${1:-$HOME/Code/brave/ethereum-remote-client}
+    profile=${2:-$HOME/Test/profiles/dev-wallet}
+    if [[ ! -d "$wallet/dist/brave" ]]; then
+        echo "No unpacked extension found in $wallet/dist/brave, did you run yarn dev:brave?"
+        return 1
+    fi
+
+    brave_profile "$brave_bin" "$profile" --load-extension="$wallet/dist/brave" -- chrome://extensions chrome://wallet
 }
 
 brave_nightly() {
@@ -32,21 +81,17 @@ brave_nightly() {
 }
 
 brave_nightly_clean() {
-    tmpdir=$(mktemp -d "/tmp/brave.XXXXXXX")
-
-    brave_nightly --user-data-dir="$tmpdir" "$@"
-    if [[ -d "$tmpdir" ]]; then
-        rm -rf "$tmpdir"
-    fi
+    brave_tmp_profile "$BRAVE_NIGHTLY"
 }
 
 brave_nightly_dev_wallet() {
+    profile=${1:-$HOME/Test/profiles/dev-wallet}
     if [[ ! -d dist/brave ]]; then
         echo "No unpacked extension found in $PWD/dist/brave, did you run yarn dev:brave?"
         return 1
     fi
 
-    brave_nightly_clean --load-extension=dist/brave brave://wallet
+    brave_profile "$BRAVE_NIGHTLY" "$profile" --load-extension=dist/brave -- chrome://extensions chrome://wallet
 }
 
 brave_ramdisk_create() {
